content_recommendation.py
```python
import random

class ContentRecommendation:
    def __init__(self, user_profile, filtered_content):
        self.user_profile = user_profile
        self.filtered_content = filtered_content

    def recommend_content(self):
        print("Generating content recommendations...")
        recommended_content = self.get_recommended_content()
        return recommended_content

    def get_recommended_content(self):
        # This is a placeholder function. You need to implement the logic to recommend content based on the user's interests and reading habits.
        # For example, you could recommend content that matches the user's interests and also fits their reading habits.
        # You could also implement a system to prioritize certain types of content based on the user's reading habits.
        # For now, this function just returns a random selection of the filtered content.
        return random.sample(self.filtered_content, min(5, len(self.filtered_content)))

    def update_user_profile(self, user_feedback):
        print("Updating user profile based on feedback...")
        # This is a placeholder function. You need to implement the logic to update the user's interests and reading habits based on their feedback.
        # For example, if the user consistently likes or dislikes certain types of content, you could update their interests or reading habits accordingly.
        # For now, this function just prints the user's feedback.
        print(user_feedback)
```
